requirements.txt
beautifulsoup4==4.12.2
openai==0.27.4
pandas==1.5.3
pdfminer.six==20221105
pdfplumber==0.8.1
pikepdf==7.2.0
PyPDF2==3.0.1
Requests==2.30.0
scikit_learn==0.24.1
streamlit==1.21.0
